print Erklaerungsfeld in Tutorial
